Implement MatchingTracker
I'm reasonably sure the strategy wll work, but as usual it's complex enough and
not automatically tested that I doubt the implementation is bug-free. The fact
that local testing has all the users with the same displayName certainly makes
it more difficult. I might put in some hacks to disambiguate display names for
local testing at least (append the playerId).

Also fill out the readme as a design doc for the automatic
matching part.