Impement AutoMatcher and friends
and uploaded the demo world. After a few short tests, it works
in the happy path, but I suspect the defect rate is still quite high;
it crashed once we got to 4 players. I think while most of the defects
are implementation errors, there are probably some design defects
lurking as well.

Particularly concerning is that networking ownership
doesn't seem to initialize before the udonbehaviors are run; joining
players kept taking ownership of the first sync object. I think I'll
need to pick sync objects by player ordinal instead of at random, which
will hopefully sync faster; given the amount of times I actually sort
the player array with selection sort it doesn't actually seem that bad.

Also got some feedback on the mechanism design itself and how it compares to
the old sdk2 system where you're locked in for 8 rounds or so of matching. Some
preferred the 'locked in' approach vs this fluid lobby system; If the teleport back
from the room is moved on to the spawn itself I think this system can get the good
parts of that same effect, while still retaining the benefits of being able to
join (or rejoin) in mid-session.

I'd like to do the manual matching map as well, with the lines on the floor;
now that I know how the player colliders work, I can do the checking without
some embarrasing N^2 distance checking between players.