Replace behind-the-head menu with double trigger, fix scrolling
Everyone complained about the pickup in their head. The double trigger
is just about as intuitive and works better. Couldn't eliminate the
update loop, but did eliminate another udonbehavior which is good.

https://booth.pm/en/items/3624917 was another option which disables
the pickup until the player hands are actually in range. heavyweight,
but could work.

also fix the scrolling. the content size fitter and unity UI is really
confusing, but I think it worgs now.